[Rising] 分离各个模型
