+++ CyxbsMobile2019_iOS/Pod/Rising Extension/UIKit Extension/RisingUIKitExtension.h
@@ -5,6 +5,7 @@
 //  Created by SSR on 2022/7/11.
 //
 
+#if __has_include(<UIKit/UIKit.h>)
 #import <UIKit/UIKit.h>
 
 #ifndef RisingUIKitExtension_h
@@ -13,6 +14,8 @@
 #import "UIView+Rising.h"
 
 #import "UIColor+Rising.h"
+
+#import <YYKit/UIImage+YYAdd.h>
 
 // UICollectionElementKind
 
@@ -25,3 +28,5 @@ FOUNDATION_EXPORT NSString *const UICollectionElementKindSectionPlaceholder API_
 FOUNDATION_EXPORT CGFloat StatusBarHeight(void);
 
 #endif /* RisingUIKitExtention_h */
+
+#endif
