+++ CyxbsMobile2019_iOS/Pod/Rising Extension/UIKit Extension/UIColor+Rising.h
@@ -5,8 +5,6 @@
 //  Created by SSR on 2022/7/28.
 //
 
-#import <UIKit/UIKit.h>
-
 #if __has_include(<YYKit/UIColor+YYAdd.h>)
 #import <YYKit/UIColor+YYAdd.h>
 
@@ -57,6 +55,5 @@ NS_ASSUME_NONNULL_BEGIN
 @end
 
 NS_ASSUME_NONNULL_END
-
 
 #endif
