+++ CyxbsMobile2019_iOS/PrefixHeader.pch
@@ -13,8 +13,4 @@
 
 #import "RisingExtension.h"
 
-#import "HttpTool.h"
-
-#import <YYKit/YYKit.h>
-
 #endif
