+++ CyxbsMobile2019_iOS/Rising Schedule/ScheduleRequestType.m
@@ -15,21 +15,3 @@ ScheduleModelRequestType const ScheduleModelRequestStudent = @"student";
 ScheduleModelRequestType const ScheduleModelRequestCustom = @"custom";
 
 ScheduleModelRequestType const ScheduleModelRequestTeacher = @"teacher";
-
-NSString *KeyInParameterForScheduleModelRequestType(ScheduleModelRequestType type) {
-    NSCParameterAssert(type);
-    
-    static NSDictionary *KeyDictionary;
-    static dispatch_once_t onceToken;
-    dispatch_once(&onceToken, ^{
-        KeyDictionary = @{
-            ScheduleModelRequestStudent : @"stu_num",
-            ScheduleModelRequestTeacher : @"tea",
-            ScheduleModelRequestCustom : @"stu_num"
-        };
-    });
-    
-    NSCParameterAssert(KeyDictionary[type]);
-    
-    return KeyDictionary[type];
-}
