+++ CyxbsMobile2019_iOS/Rising Schedule/ScheduleServiceDataSource.h
@@ -19,6 +19,8 @@
 
 #import "ScheduleCollectionViewLayout.h"
 
+#import "ScheduleCollectionViewDataSource.h"
+
 NS_ASSUME_NONNULL_BEGIN
 
 #pragma mark - ScheduleServiceDataSource
